chore: deprecate cryptutils package
This commit deprecates the `cryptutils` package.
