--- cryptutils/mode_cfb.go
@@ -1,12 +1,0 @@
-package cryptutils
-
-import (
-	"crypto/cipher"
-)
-
-func CFB(block cipher.Block, iv []byte) CryptStream {
-	return streamBlock{
-		enc: cipher.NewCFBEncrypter(block, iv),
-		dec: cipher.NewCFBDecrypter(block, iv),
-	}
-}
