--- cryptutils/mode_gcm.go
@@ -1,9 +1,0 @@
-package cryptutils
-
-import (
-	"crypto/cipher"
-)
-
-func GCM(block cipher.Block) (AEAD, error) {
-	return cipher.NewGCM(block)
-}
